requirements.txt
pandas
streamlit
google-generativeai
PyPDF2
langchain
